requirements.txt
alabaster==0.7.10
Babel==2.5.0
backports.weakref==1.0rc1
bleach==1.5.0
cairocffi==0.8.0
CairoSVG==2.0.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
contextlib2==0.5.5
cssselect==1.0.1
cycler==0.10.0
distro==1.0.4
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.11.1
django-admin-interface==0.5.1
django-auth0==0.0.3
django-bootstrap-breadcrumbs==0.8.2
django-colorfield==0.1.14
django-crispy-forms==1.6.1
django-developer-panel==0.1.2
django-filter==1.0.4
django-flat-responsive==1.4.1
django-flat-theme==1.1.4
django-progressive-web-app==0.1
django-report-builder==3.5.0
djangorestframework==3.6.3
dlib==19.4.0
docutils==0.14
et-xmlfile==1.0.1
face-recognition==0.2.2
face-recognition-models==0.1.3
gunicorn==19.6.0
h5py==2.7.0
html5lib==0.999999999
idna==2.5
imagesize==0.7.1
jdcal==1.3
Jinja2==2.9.6
lxml==3.8.0
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.0.2
numpy==1.13.1
olefile==0.44
opencv-contrib-python==3.2.0.8
opencv-python==3.2.0.8
openpyxl==2.4.8
Pillow==4.0.0
protobuf==3.3.0
psutil==5.2.2
psycopg2==2.6.2
pycparser==2.18
Pygments==2.2.0
PyInstaller==3.2.1
pyparsing==2.2.0
Pyphen==0.9.4
PyQt5==5.9
python-dateutil==2.6.1
pytz==2017.2
pywebview==1.7
raven==6.1.0
requests==2.18.3
scikit-learn==0.18.2
scipy==0.19.1
sip==4.19.3
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
static3==0.7.0
tensorflow==1.2.1
tinycss==0.4
tinycss2==0.5
urllib3==1.22
WeasyPrint==0.39
webencodings==0.5.1
Werkzeug==0.12.2
whitenoise==3.2
